Fix the problem of the game not closing
Previously, due to the fact that first the window was closed and
only then the statestack was removed this macro glGetError constantly
returned GL_INVALID_OPERATION. This led to an infinite loop.

Now, before the window is closed the StateStack is forced to remove
all the state by which buffers and shaders are removed before the
window is closed.