Add player physics
It makes the player collide with other objects in the game.
Such a collider must have the tag set to solid.

DrawableCollider has a changed constructor due to the fact
that DrawableCollider's copy-constructor previously did not
work well, as the child class object was passed to the collider's
constructor. Now such a constructor is excluded from the search.